+++ examples/HybridRrf/Data/MockItemWithEmbedding.cs
@@ -1,9 +1,8 @@
-using Pgvector;
 using Shared;
 
 namespace HybridRrf.Data;
 
 public class MockItemWithEmbedding : MockItem
 {
-    public Vector? Embedding { get; set; }
+    public float[]? Embedding { get; set; }
 }